user login

+++ app/Http/Controllers/TasksController.php
@@ -96,9 +96,13 @@ class TasksController extends Controller
     {
         $task = Task::find($id);
 
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+        }
+        
+        return redirect('/');
     }
 
     /**
@@ -116,10 +120,15 @@ class TasksController extends Controller
         ]);
 
         $task = Task::find($id);
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
 
+        if (\Auth::id() === $task->user_id) {
+            $task->status = $request->status;
+            $task->content = $request->content;
+            $task->save();
+    
+            return redirect('/');
+        }
+        
         return redirect('/');
     }
 
@@ -132,8 +141,13 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
-        $task->delete();
+        
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
 
+            return redirect('/');
+        }
+        
         return redirect('/');
     }
 }
